Fix exports from GitHub Enterprise Server, and validate the version when starting an export
In https://github.com/timrogers/gh-migrate-project/discussions/67,
@harvi2048 spotted that exports from GitHub Enterprise Server
didn't work (in their case from 3.9.5) because we try to query
some fields that don't exist in current GitHub Enterprise Server
versions - namely, the `description` and `color` on "Single select"
custom fields' options.

This fixes that by checking what environment we are exporting from
at the beginning of the process, and then conditionally including
or excluding those fields from the GraphQL query based on the
version.

As part of this change, we also check that we are exporting from
a supported GitHub product - at this time, either GitHub.com or
any GitHub Enterprise Server version with Projects (that is, 3.7+).