Explicitly only support imports to GitHub.com
This tool was originally designed to allow exports and imports
from/to GitHub.com and GitHub Enterprise Server, but this wasn't
fully tested, and the reality is complicated!

Imports to GitHub Enterprise Server are not working properly, and
fixing this will require some work due to API differences.

Instead, this change makes the situation clear: only imports to
GitHub.com are supported, although we will endeavour to accept
exports from any GHES version - which means, for example, filling
in some data that has been mandatory since older versions.

As a follow up, we can look at fixing support for GHES imports.