Avoid outputting GitHub tokens from environment variables in CLI help output
When calling any of our CLI commands with `--help`, if the GitHub
access token is set using an environment variable, its value is
included in the help output in plain text. This isn't desirable.

This changes our implementation of "environment variable as
default" so that the environment variable is still used, but it
isn't displayed as the default in help output.