When an error is caught at the command level, log its stack trace
Each of our commands' handlers is wrapped with an error handler,
`actionErrorHandler`, which logs the error message in red and
then exits with status code `1`.

An error's message is not necessarily super helpful - as seen
with the `Cannot read properties of null (reading '__typename')`
error in #85. This additionally logs the stack trace, giving more
context on what went wrong and where.