Check for updates to the extension when running the `export` or `import` commands
This adds a new automated update check when the extension is run.

Unless the `--skip-update-check` flag is set, we'll fetch the
latest extension version using the GitHub API and check that
you're up to date.

If you're running an old version, we'll let you know what the
latest version is and explain how to update.