Gracefully handle missing a missing issue/PR when importing a project item
When trying to import a project item where the pull request or
issue is missing, this is intended to log a warning, but currently,
it causes the entire application to crash.

This fixes it, so a warning is logged and we move on to importing
the next project item.

Fixes #227.