feat: support running imports targeting an already-created project
As proposed by @cameronraysmith in #395, this adds the ability for
the `import` command to target an existing project that has already
been created. The import process will take that project, setup the
fields and add the items.

This can be useful for circumventing the GitHub API's lack of
support for managing project workflows and views, which means that
this tool can't support migrating that data. Instead, you can set
up those things in a project template, create a project from that
template, and then run the import.

Closes #395.